game time
turned it into more of a proper game, with ui elements such as player
health, score, and a progress bar for the group hug. also added particle
effects for the group hug and the blow kiss mechanic. the AI is still
buggy as hell, but I added a script to fix one of the bugs where it
would randomly fall through the floor.